convert literals to units
